Moved text command manager, added declaration interface
