+++ src/main/kotlin/io/github/freya022/botcommands/api/core/service/annotations/InterfacedService.kt
@@ -9,6 +9,7 @@ import io.github.freya022.botcommands.api.commands.text.HelpBuilderConsumer
 import io.github.freya022.botcommands.api.commands.text.IHelpCommand
 import io.github.freya022.botcommands.api.commands.text.TextCommandFilter
 import io.github.freya022.botcommands.api.commands.text.TextSuggestionSupplier
+import io.github.freya022.botcommands.api.commands.text.declaration.TextCommandsDeclaration
 import io.github.freya022.botcommands.api.components.ComponentInteractionFilter
 import io.github.freya022.botcommands.api.core.*
 import io.github.freya022.botcommands.api.core.config.BServiceConfigBuilder
@@ -43,6 +44,7 @@ import io.github.freya022.botcommands.api.localization.readers.LocalizationMapRe
  *
  * @see GlobalExceptionHandler
  *
+ * @see TextCommandsDeclaration
  * @see TextCommandFilter
  * @see DefaultEmbedSupplier
  * @see DefaultEmbedFooterIconSupplier
