+++ src/test/kotlin/io/github/freya022/botcommands/test/commands/text/TextFilter.kt
@@ -3,10 +3,14 @@ package io.github.freya022.botcommands.test.commands.text
 import dev.minn.jda.ktx.coroutines.await
 import io.github.freya022.botcommands.api.commands.annotations.Command
 import io.github.freya022.botcommands.api.commands.annotations.Filter
-import io.github.freya022.botcommands.api.commands.text.*
+import io.github.freya022.botcommands.api.commands.text.BaseCommandEvent
+import io.github.freya022.botcommands.api.commands.text.TextCommand
+import io.github.freya022.botcommands.api.commands.text.and
 import io.github.freya022.botcommands.api.commands.text.annotations.JDATextCommandVariation
 import io.github.freya022.botcommands.api.commands.text.annotations.TextDeclaration
 import io.github.freya022.botcommands.api.commands.text.builder.filter
+import io.github.freya022.botcommands.api.commands.text.declaration.TextCommandManager
+import io.github.freya022.botcommands.api.commands.text.or
 import io.github.freya022.botcommands.test.filters.InVoiceChannel
 import io.github.freya022.botcommands.test.filters.IsBotOwner
 import io.github.freya022.botcommands.test.filters.IsGuildOwner
