+++ src/test/kotlin/io/github/freya022/botcommands/test/commands/text/TextVararg.kt
@@ -3,8 +3,8 @@ package io.github.freya022.botcommands.test.commands.text
 import io.github.freya022.botcommands.api.annotations.CommandMarker
 import io.github.freya022.botcommands.api.commands.annotations.Command
 import io.github.freya022.botcommands.api.commands.text.BaseCommandEvent
-import io.github.freya022.botcommands.api.commands.text.TextCommandManager
 import io.github.freya022.botcommands.api.commands.text.annotations.TextDeclaration
+import io.github.freya022.botcommands.api.commands.text.declaration.TextCommandManager
 
 @Command
 class TextVararg {
